feat: 도서 번호로 리뷰 조회, html 파일 디렉토리 이동

+++ src/main/java/com/nhnacademy/inkbridge/front/dto/review/ReviewAverageReadResponseDto.java
@@ -1,6 +1,5 @@
 package com.nhnacademy.inkbridge.front.dto.review;
 
-import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -14,6 +13,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ReviewAverageReadResponseDto {
 
-    private BigDecimal avg;
+    private Double avg;
     private Long count;
 }
